+++ nkululeko/constants.py
@@ -1,4 +1,4 @@
-VERSION = "1.8.14"
+VERSION = "1.9.0"
 SAMPLING_RATE = 16000
 COL_SEX = "gender"
 COL_AGE = "age"